This PR multithreads all resize calls which are the slowest part of mosaic that can be ran in parallel.
Experimental, needs a bit more testing.

# Benchmark 1
Before:
```
Took 268ms (mosaic: 108ms, encoding: 160ms)
Took 254ms (mosaic: 98ms, encoding: 155ms) 
Took 249ms (mosaic: 102ms, encoding: 147ms)
Took 262ms (mosaic: 105ms, encoding: 156ms)
Took 266ms (mosaic: 108ms, encoding: 158ms)
Took 267ms (mosaic: 103ms, encoding: 163ms)
```

After:
```
Took 216ms (mosaic: 59ms, encoding: 157ms)
Took 229ms (mosaic: 55ms, encoding: 174ms)
Took 222ms (mosaic: 67ms, encoding: 154ms)
Took 219ms (mosaic: 61ms, encoding: 157ms)
Took 222ms (mosaic: 54ms, encoding: 168ms)
Took 233ms (mosaic: 57ms, encoding: 175ms)
Took 200ms (mosaic: 54ms, encoding: 145ms)
Took 212ms (mosaic: 58ms, encoding: 153ms)
Took 214ms (mosaic: 54ms, encoding: 160ms)
```

Image:
![image](https://user-images.githubusercontent.com/32272859/184464112-161613e0-3586-47e6-8131-3c8930b95bfe.png)

# Benchmark 2
Before:
```
Took 1481ms (mosaic: 745ms, encoding: 735ms)
Took 1570ms (mosaic: 854ms, encoding: 716ms)
Took 1700ms (mosaic: 886ms, encoding: 813ms)
Took 1701ms (mosaic: 950ms, encoding: 750ms)
Took 1601ms (mosaic: 859ms, encoding: 742ms)
```
After:
```
Took 1161ms (mosaic: 409ms, encoding: 752ms)
Took 1182ms (mosaic: 385ms, encoding: 797ms)
Took 1145ms (mosaic: 383ms, encoding: 761ms)
Took 1263ms (mosaic: 477ms, encoding: 786ms)
Took 1274ms (mosaic: 458ms, encoding: 816ms)
```
Image:
![image](https://user-images.githubusercontent.com/32272859/184464122-dca48e3c-b02c-47b3-9db5-0e4759b3cfe9.png)